最热文章

+++ article/views.py
@@ -18,8 +18,15 @@ from django.core.paginator import Paginator
 
 # 文章列表
 def article_list(request):
-    # 取出所有博客文章
-    article_list = ArticlePost.objects.all()
+    # 根据GET请求中查询条件
+    # 返回不同排序的对象数组
+    if request.GET.get('order') == 'total_views':
+        article_list = ArticlePost.objects.all().order_by('-total_views')
+        order = 'total_views'
+    else:
+        # 取出所有博客文章
+        article_list = ArticlePost.objects.all()
+        order = 'normal'
     # 每页显示 1 篇文章
     paginator = Paginator(article_list, 3)
     # 获取 url 中的页码
@@ -27,7 +34,7 @@ def article_list(request):
     # 将导航对象相应的页码内容返回给 articles
     articles = paginator.get_page(page)
     # 需要传递给模板（templates）的对象
-    context = { 'articles': articles }
+    context = { 'articles': articles, 'order': order }
     # render函数：载入模板，并返回context对象
     return render(request, 'article/list.html', context)
 
@@ -87,9 +94,13 @@ def article_create(request):
 
 
 # 删除文章
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
     # 根据 id 获取需要删除的文章
     article = ArticlePost.objects.get(id=id)
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     # 调用.delete()方法删除文章
     article.delete()
     # 完成删除后返回文章列表
@@ -97,6 +108,7 @@ def article_delete(request, id):
 
 
 # 更新文章
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -107,6 +119,11 @@ def article_update(request, id):
 
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
+
     # 判断用户是否为 POST 提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
